Use new unwinder for offline in libbacktrace.
libunwindstack changes:
- Add a load_bias field in MapInfo so that it can be loaded offline,
  and also so it can be cached.
- Add an Add function the Maps class so that it's possible to manually
  create a map.
- Remove the OfflineMaps class since I haven't found a reason for this to
  exist.
- Add a pointer to the gnu debugdata compressed section in the interface
  itself and modify the step path to try eh_frame, then debug_frame, then
  gnu_debugdata. This way arm can add exidx as the last step behind
  gnu_debugdata. Add an offline test for this feature.
- Fix x86_64_ucontext_t since it was a different size on 32 bit and 64 bit
  systems.

libbbacktrace changes:
- Completely rewrite the BacktraceOffline class to use the new unwinder.
- Modify the test data to save ucontext_t data instead of unw_context data.
- Convert the previous tests from unw_context data to ucontext_t data.
- Change uintptr_t to uint64_t to allow offline unwinds to work properly.

debuggerd changes:
- Change some uintptr_t to uint64_t to correspond with libbacktrace changes.

Bug: 65682279

Test: New unit tests pass in libunwindstack.
Test: All offline tests continue to pass.

+++ debuggerd/libdebuggerd/include/libdebuggerd/utility.h
@@ -64,7 +64,7 @@ void _LOG(log_t* log, logtype ltype, const char *fmt, ...)
 
 bool wait_for_signal(pid_t tid, siginfo_t* siginfo);
 
-void dump_memory(log_t* log, Backtrace* backtrace, uintptr_t addr, const char* fmt, ...);
+void dump_memory(log_t* log, Backtrace* backtrace, uint64_t addr, const char* fmt, ...);
 
 void read_with_default(const char* path, char* buf, size_t len, const char* default_value);
